requirements.txt
fastapi[standard]==0.127.0
alembic==1.17.2
SQLAlchemy==2.0.45
psycopg2-binary==2.9.11
pwdlib[argon2]==0.3.0
pyjwt[crypto]==2.10.1

